update in store and item taxes

+++ controllers/stores.js
@@ -2,6 +2,7 @@ import express from "express";
 import Store from "../modals/Store";
 import POS_Device from "../modals/POS_Device";
 import diningOption from "../modals/settings/diningOption";
+
 const router = express.Router();
 
 router.post("/", async (req, res) => {
@@ -48,7 +49,11 @@ router.post("/", async (req, res) => {
         console.log("Default Dining Insert Error", err.message);
       });
   } catch (error) {
-    res.status(400).json({ message: error.message });
+    if (error.code === 11000) {
+      res.status(400).json({ message: "Store Already Register By This User" });
+    } else {
+      res.status(400).json({ message: error.message });
+    }
   }
 });
 router.get("/", async (req, res) => {
@@ -81,12 +86,35 @@ router.get("/", async (req, res) => {
 router.delete("/:ids", async (req, res) => {
   try {
     var { ids } = req.params;
+    let message = [];
     ids = JSON.parse(ids);
-    ids.forEach(async (id) => {
-      await Store.deleteOne({ _id: id });
-    });
+    // ids.forEach(async (id) => {
+    for (const id of ids) {
+      let devices = await POS_Device.find({
+        "store.storeId": id,
+      }).countDocuments();
+      let result = await Store.findOne({ _id: id }).sort({ _id: "desc" });
+      if (devices === 0) {
+        await Store.deleteOne({ _id: id });
+        if (result !== undefined || result !== null) {
+          message.push({
+            message: `This Store (${result.title}) Is Deleted Successfully`,
+            error: false,
+            storeId: id,
+          });
+        }
+      } else {
+        if (result !== undefined || result !== null) {
+          message.push({
+            message: `This Store (${result.title}) Use In POS Device`,
+            error: true,
+            storeId: id,
+          });
+        }
+      }
+    }
 
-    res.status(200).json({ message: "deleted" });
+    res.status(200).json({ message: message });
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
